Delete most remnants of references to GBA PPU registers Reimplement scanline effects

+++ include/battle.h
@@ -609,18 +609,18 @@ struct MonSpritesGfx
 };
 
 // All battle variables are declared in battle_main.c
-extern u16 gBattle_BG0_X;
-extern u16 gBattle_BG0_Y;
-extern u16 gBattle_BG1_X;
-extern u16 gBattle_BG1_Y;
-extern u16 gBattle_BG2_X;
-extern u16 gBattle_BG2_Y;
-extern u16 gBattle_BG3_X;
-extern u16 gBattle_BG3_Y;
-extern u16 gBattle_WIN0H;
-extern u16 gBattle_WIN0V;
-extern u16 gBattle_WIN1H;
-extern u16 gBattle_WIN1V;
+extern u32 gBattle_BG0_X;
+extern u32 gBattle_BG0_Y;
+extern u32 gBattle_BG1_X;
+extern u32 gBattle_BG1_Y;
+extern u32 gBattle_BG2_X;
+extern u32 gBattle_BG2_Y;
+extern u32 gBattle_BG3_X;
+extern u32 gBattle_BG3_Y;
+extern u32 gBattle_WIN0H;
+extern u32 gBattle_WIN0V;
+extern u32 gBattle_WIN1H;
+extern u32 gBattle_WIN1V;
 extern u8 gDisplayedStringBattle[300];
 extern u8 gBattleTextBuff1[TEXT_BUFF_ARRAY_COUNT];
 extern u8 gBattleTextBuff2[TEXT_BUFF_ARRAY_COUNT];
